+++ dataforge-data/src/commonMain/kotlin/space/kscience/dataforge/data/DataSetBuilder.kt
@@ -2,11 +2,11 @@ package space.kscience.dataforge.data
 
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.meta.MutableMeta
 import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.dataforge.names.Name
+import space.kscience.dataforge.names.isEmpty
 import space.kscience.dataforge.names.plus
 import kotlin.reflect.KType
 
@@ -18,12 +18,12 @@ public interface DataSetBuilder<in T : Any> {
      */
     public suspend fun remove(name: Name)
 
-    public suspend fun emit(name: Name, data: Data<T>?)
+    public suspend fun data(name: Name, data: Data<T>?)
 
     /**
      * Set a current state of given [dataSet] into a branch [name]. Does not propagate updates
      */
-    public suspend fun emit(name: Name, dataSet: DataSet<T>) {
+    public suspend fun node(name: Name, dataSet: DataSet<T>) {
         //remove previous items
         if (name != Name.EMPTY) {
             remove(name)
@@ -31,27 +31,29 @@ public interface DataSetBuilder<in T : Any> {
 
         //Set new items
         dataSet.flowData().collect {
-            emit(name + it.name, it.data)
+            data(name + it.name, it.data)
         }
     }
 
     /**
-     * Append data to node
+     * Set meta for the given node
      */
-    public suspend infix fun String.put(data: Data<T>): Unit = emit(Name.parse(this), data)
+    public suspend fun meta(name: Name, meta: Meta)
 
-    /**
-     * Append node
-     */
-    public suspend infix fun String.put(dataSet: DataSet<T>): Unit = emit(Name.parse(this), dataSet)
-
-    /**
-     * Build and append node
-     */
-    public suspend infix fun String.put(block: suspend DataSetBuilder<T>.() -> Unit): Unit = emit(Name.parse(this), block)
 }
 
-private class SubSetBuilder<in T : Any>(
+/**
+ * Define meta in this [DataSet]
+ */
+public suspend fun <T : Any> DataSetBuilder<T>.meta(value: Meta): Unit = meta(Name.EMPTY, value)
+
+/**
+ * Define meta in this [DataSet]
+ */
+public suspend fun <T : Any> DataSetBuilder<T>.meta(mutableMeta: MutableMeta.() -> Unit): Unit = meta(Meta(mutableMeta))
+
+@PublishedApi
+internal class SubSetBuilder<in T : Any>(
     private val parent: DataSetBuilder<T>,
     private val branch: Name,
 ) : DataSetBuilder<T> {
@@ -61,33 +63,42 @@ private class SubSetBuilder<in T : Any>(
         parent.remove(branch + name)
     }
 
-    override suspend fun emit(name: Name, data: Data<T>?) {
-        parent.emit(branch + name, data)
+    override suspend fun data(name: Name, data: Data<T>?) {
+        parent.data(branch + name, data)
     }
 
-    override suspend fun emit(name: Name, dataSet: DataSet<T>) {
-        parent.emit(branch + name, dataSet)
+    override suspend fun node(name: Name, dataSet: DataSet<T>) {
+        parent.node(branch + name, dataSet)
+    }
+
+    override suspend fun meta(name: Name, meta: Meta) {
+        parent.meta(branch + name, meta)
     }
 }
 
-public suspend fun <T : Any> DataSetBuilder<T>.emit(name: Name, block: suspend DataSetBuilder<T>.() -> Unit) {
-    SubSetBuilder(this, name).apply { block() }
+public suspend inline fun <T : Any> DataSetBuilder<T>.node(
+    name: Name,
+    crossinline block: suspend DataSetBuilder<T>.() -> Unit,
+) {
+    if (name.isEmpty()) block() else SubSetBuilder(this, name).block()
 }
 
 
-public suspend fun <T : Any> DataSetBuilder<T>.emit(name: String, data: Data<T>) {
-    emit(Name.parse(name), data)
+public suspend fun <T : Any> DataSetBuilder<T>.data(name: String, value: Data<T>) {
+    data(Name.parse(name), value)
 }
 
-public suspend fun <T : Any> DataSetBuilder<T>.emit(name: String, set: DataSet<T>) {
-    this.emit(Name.parse(name), set)
+public suspend fun <T : Any> DataSetBuilder<T>.node(name: String, set: DataSet<T>) {
+    node(Name.parse(name), set)
 }
 
-public suspend fun <T : Any> DataSetBuilder<T>.emit(name: String, block: suspend DataSetBuilder<T>.() -> Unit): Unit =
-    this@emit.emit(Name.parse(name), block)
+public suspend inline fun <T : Any> DataSetBuilder<T>.node(
+    name: String,
+    crossinline block: suspend DataSetBuilder<T>.() -> Unit,
+): Unit = node(Name.parse(name), block)
 
-public suspend fun <T : Any> DataSetBuilder<T>.emit(data: NamedData<T>) {
-    emit(data.name, data.data)
+public suspend fun <T : Any> DataSetBuilder<T>.set(value: NamedData<T>) {
+    data(value.name, value.data)
 }
 
 /**
@@ -99,7 +110,7 @@ public suspend inline fun <reified T : Any> DataSetBuilder<T>.produce(
     noinline producer: suspend () -> T,
 ) {
     val data = Data(meta, block = producer)
-    emit(name, data)
+    data(name, data)
 }
 
 public suspend inline fun <reified T : Any> DataSetBuilder<T>.produce(
@@ -108,7 +119,7 @@ public suspend inline fun <reified T : Any> DataSetBuilder<T>.produce(
     noinline producer: suspend () -> T,
 ) {
     val data = Data(meta, block = producer)
-    emit(name, data)
+    data(name, data)
 }
 
 /**
@@ -117,36 +128,34 @@ public suspend inline fun <reified T : Any> DataSetBuilder<T>.produce(
 public suspend inline fun <reified T : Any> DataSetBuilder<T>.static(
     name: String,
     data: T,
-    meta: Meta = Meta.EMPTY
-): Unit =
-    emit(name, Data.static(data, meta))
+    meta: Meta = Meta.EMPTY,
+): Unit = data(name, Data.static(data, meta))
 
 public suspend inline fun <reified T : Any> DataSetBuilder<T>.static(
     name: Name,
     data: T,
-    meta: Meta = Meta.EMPTY
-): Unit =
-    emit(name, Data.static(data, meta))
+    meta: Meta = Meta.EMPTY,
+): Unit = data(name, Data.static(data, meta))
 
 public suspend inline fun <reified T : Any> DataSetBuilder<T>.static(
     name: String,
     data: T,
     mutableMeta: MutableMeta.() -> Unit,
-): Unit = emit(Name.parse(name), Data.static(data, Meta(mutableMeta)))
+): Unit = data(Name.parse(name), Data.static(data, Meta(mutableMeta)))
 
 /**
  * Update data with given node data and meta with node meta.
  */
 @DFExperimental
-public suspend fun <T : Any> DataSetBuilder<T>.populate(tree: DataSet<T>): Unit = coroutineScope {
+public suspend fun <T : Any> DataSetBuilder<T>.populateFrom(tree: DataSet<T>): Unit = coroutineScope {
     tree.flowData().collect {
         //TODO check if the place is occupied
-        emit(it.name, it.data)
+        data(it.name, it.data)
     }
 }
 
-public suspend fun <T : Any> DataSetBuilder<T>.populate(flow: Flow<NamedData<T>>) {
+public suspend fun <T : Any> DataSetBuilder<T>.populateWith(flow: Flow<NamedData<T>>) {
     flow.collect {
-        emit(it.name, it.data)
+        data(it.name, it.data)
     }
 }